# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Converto Business OS API
  version: 1.0.0
  description: DocFlow backend API for document processing, admin management, and business operations
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.docflow.fi
    description: Production
paths:
  /api/admin/activities:
    get:
      operationId: getAdminActivities
      summary: Get admin activities
      tags:
        - Admin
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: tenant_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
    post:
      operationId: createAdminActivity
      summary: Create admin activity
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                details:
                  type: object
                tenant_id:
                  type: string
      responses:
        '201':
          description: Activity created
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
  /api/admin/feed:
    get:
      operationId: adminWebSocketFeed
      summary: Admin WebSocket feed
      tags:
        - Admin
      description: WebSocket endpoint for real-time activity feed
      responses:
        '101':
          description: Switching protocols to WebSocket
      security:
        - BearerAuth: []
...
components:
  schemas:
    ActivityResponse:
      type: object
      properties:
        activities:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT